feat: 파일 최대 용량 설정
(cherry picked from commit cfb7bdcf97bff518dc1e388c4d1bd61e2d0224a8)

+++ src/main/java/application/member/domain/Membership.java
@@ -22,7 +22,7 @@ public class Membership {
     }
 
     public void uploadFile(long fileSize) {
-        int savedFileLimitMegaByte = membershipType.getSavedFileLimitByte();
+        long savedFileLimitMegaByte = membershipType.getSavedFileLimitByte();
         if (savedFileLimitMegaByte < currentSavedFileBytes + fileSize) {
             throw new ForbiddenException("더이상 파일을 저장할 수 없습니다.");
         }
